fix: 채팅방 포화도 컬럼 타입 변경
실수 float 타입지정 시 범위 연산이 정확하지 않은 문제
포화도 수치를 실수로 저장하던 방식을 백분율로 수정
소수점 단위 버림

+++ module-api/src/main/resources/sql/test/schema.sql
@@ -21,7 +21,7 @@ CREATE TABLE `Room` (
     password VARCHAR(255),
     capacity INTEGER NOT NULL,
     attending INTEGER NOT NULL,
-    saturation FLOAT NOT NULL,
+    saturation INTEGER NOT NULL,
     createdAt TIMESTAMP(3) NOT NULL,
     removedAt TIMESTAMP(3),
     PRIMARY KEY (id)
